requirements.txt
requests
bs4
lxml
langchain-community
duckduckgo-search